stake management refactor - rewards code reorg
